Refactor solution.py to remove unneeded array
A previous verions of this script used an asteroid object and a loop
in solution.py created all the asteroid objects. This was subsequently
streamlined and the asteroid class removed in favor of an array, but
the loop remained even though all it was doing was copying one array to
another. Now it has been removed.